requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0
openai==1.3.0
boto3==1.34.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
Pillow==10.0.1
gunicorn==21.2.0
deepgram-sdk==2.12.0 